Move to use templating for SQL.
I'm pretty sick of constructing SQL statements in python.  So I wrote an extremely simple library (https://github.com/owainkenwayucl/simpletemplate) which essentially search and replaces on a file with a dictionary of key/value pairs and returns it as a string.

This means I can split the SQL queries out into their own SQL files with some arbitrary keys in them (%KEY% in this case) to use as templates.

I've included a copy of my simple library in the source tree of this project so it doesn't break anything for anyone.